+++ nas/models/blocks/SuperResConvK1KXK1.py
@@ -24,6 +24,8 @@ class ResConvK1KXK1(nn.Module):
             'p': padding (default=(k-1)//2,
             'g': grouping (default=1),
             'btn': bottleneck_channels,
+            'nbitsA': input activation quant nbits list(default=8),
+            'nbitsW': weight quant nbits list(default=8),
             'act': activation (default=relu),
         }
         :param NAS_mode:
@@ -53,6 +55,16 @@ class ResConvK1KXK1(nn.Module):
         else:
             self.force_resproj = False
 
+        if "nbitsA" in structure_info and "nbitsW" in structure_info:
+            self.quant = True
+            self.nbitsA = structure_info["nbitsA"]
+            self.nbitsW = structure_info["nbitsW"]
+            if len(self.nbitsA)!=3 or len(self.nbitsW)!=3:
+                raise ValueError("nbitsA/W must has three elements in %s, not nbitsA %d or nbitsW %d"%
+                        (self.__class__, len(self.nbitsA), len(self.nbitsW)))
+        else:
+            self.quant = False
+
         if 'g' in structure_info:
             self.groups = structure_info['g']
         else:
@@ -74,6 +86,10 @@ class ResConvK1KXK1(nn.Module):
                             's': self.stride, 'g': self.groups, 'p': self.padding}
         conv3_info = {'in': self.bottleneck_channels, 'out': self.out_channels, 'k': 1,
                             's': 1, 'g': self.groups, 'p': 0}
+        if self.quant:
+            conv1_info = {**conv1_info, **{"nbitsA":self.nbitsA[0], "nbitsW":self.nbitsW[0]}}
+            conv2_info = {**conv2_info, **{"nbitsA":self.nbitsA[1], "nbitsW":self.nbitsW[1]}}
+            conv3_info = {**conv3_info, **{"nbitsA":self.nbitsA[2], "nbitsW":self.nbitsW[2]}}
 
         self.conv1 = ConvKXBN(conv1_info, no_create=no_create, **kwargs)
         self.conv2 = ConvKXBN(conv2_info, no_create=no_create, **kwargs)
@@ -106,7 +122,11 @@ class ResConvK1KXK1(nn.Module):
                 self.residual_downsample = nn.Identity()
 
         if self.in_channels != self.out_channels or self.force_resproj:
-            self.residual_proj = ConvKXBN({'in': self.in_channels, 'out': self.out_channels, 'k': 1,
+            if self.quant:
+                self.residual_proj = ConvKXBN({'in': self.in_channels, 'out': self.out_channels, 'k': 1,
+                                           's': 1, 'g': 1, 'p': 0, "nbitsA":self.nbitsA[0], "nbitsW":self.nbitsW[0]}, no_create=no_create)
+            else:
+                self.residual_proj = ConvKXBN({'in': self.in_channels, 'out': self.out_channels, 'k': 1,
                                            's': 1, 'g': 1, 'p': 0}, no_create=no_create)
             self.model_size = self.model_size + self.residual_proj.get_model_size()
             self.flops = self.flops + self.residual_proj.get_flops(1.0 / self.stride) + self.out_channels / self.stride ** 2
@@ -154,8 +174,11 @@ class ResConvK1KXK1(nn.Module):
         return output
         
 
-    def get_model_size(self):
-        return self.model_size
+    def get_model_size(self, return_list=False):
+        if return_list:
+            return self.conv1.get_model_size(return_list)+self.conv2.get_model_size(return_list)+self.conv3.get_model_size(return_list)
+        else:
+            return self.model_size
 
 
     def get_flops(self, resolution):
@@ -200,14 +223,44 @@ class ResConvK1KXK1(nn.Module):
             if not skip_relu: output = self.activation_function(output)
             # print("output std: mean %.4f, std %.4f, max %.4f, min %.4f\n"%(
                     # output.mean().item(), output.std().item(), output.max().item(), output.min().item()))
-            output_std_block *= output.std()
-            output = output/output.std()
+            if "init_std_act" in kwarg and hasattr(self, "nbitsA"):
+                output_std_block *= output.std()/kwarg["init_std_act"]
+                output = output/(output.std()/kwarg["init_std_act"])
+            else:
+                output_std_block *= output.std()
+                output = output/output.std()
         output_std_list.append(output_std_block)
         return output, output_std_list
 
 
     def get_num_channels_list(self):
         return [self.bottleneck_channels, self.bottleneck_channels, self.out_channels]
+
+
+    def get_log_zen_score(self, **kwarg):
+        if "init_std" in kwarg and "init_std_act" in kwarg and hasattr(self, "nbitsA"):
+            conv1_std = np.log(STD_BITS_LUT[kwarg["init_std_act"]][self.nbitsA[0]]*STD_BITS_LUT[kwarg["init_std"]][self.nbitsW[0]])-np.log(kwarg["init_std_act"])
+            conv2_std = np.log(STD_BITS_LUT[kwarg["init_std_act"]][self.nbitsA[1]]*STD_BITS_LUT[kwarg["init_std"]][self.nbitsW[1]])-np.log(kwarg["init_std_act"])
+            conv3_std = np.log(STD_BITS_LUT[kwarg["init_std_act"]][self.nbitsA[2]]*STD_BITS_LUT[kwarg["init_std"]][self.nbitsW[2]])-np.log(kwarg["init_std_act"])
+
+            return [np.log(np.sqrt(self.in_channels)) + conv1_std + \
+                    np.log(np.sqrt(self.bottleneck_channels * self.kernel_size ** 2)) + conv2_std + \
+                    np.log(np.sqrt(self.bottleneck_channels))+conv3_std]
+        else:
+            return [np.log(np.sqrt(self.in_channels)) + \
+                    np.log(np.sqrt(self.bottleneck_channels * self.kernel_size ** 2)) + \
+                    np.log(np.sqrt(self.bottleneck_channels))]
+
+    def get_max_feature_num(self, resolution):
+        residual_featmap = resolution**2*self.out_channels//(self.stride**2)
+        if self.quant:
+            residual_featmap = residual_featmap * self.nbitsA[0] / 8
+        conv1_max_featmap = self.conv1.get_max_feature_num(resolution) + residual_featmap
+        conv2_max_featmap = self.conv2.get_max_feature_num(resolution) + residual_featmap
+        conv3_max_featmap = self.conv3.get_max_feature_num(resolution//self.stride)
+        max_featmap_list = [conv1_max_featmap, conv2_max_featmap, conv3_max_featmap]
+
+        return max_featmap_list
 
 
 class SuperResConvK1KXK1(BaseSuperBlock):