+++ nas/spaces/space_K1KXK1.py
@@ -1,3 +1,5 @@
+# Copyright (c) 2021-2022 Alibaba Group Holding Limited.
+
 import os, sys
 import random, copy, bisect