+++ docs/changelog.md
@@ -1,8 +1,24 @@
 ## Changelog
-***
-**1.0.0** was released in 28/03/2022:
 
-* Support two scores for zero-shot search.
+
+***
+**1.0.2** was released in 2022/09/19:
+
+* Support madnas score for training-free search.
+* Support mixed-precision Quantization tasks.
+* Update the searched mixed-precision models.
+
+
+***
+**1.0.1** was released in 2022/06/08:
+
+* Update the MAE-Det searched models.
+* Update the README for the projects.
+
+
+***
+**1.0.0** was released in 2022/04/01:
+
+* Support entropy score for training-free search.
 * Support latency prediction for hardware device.
 * Support Classification and Detection tasks.
-***